@@ -4,8 +4,11 @@ import android.content.Intent;
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.design.widget.FloatingActionButton;
 import android.support.v4.view.ViewCompat;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.CardView;
+import android.support.v7.widget.Toolbar;
 import android.view.View;
 
 import com.harris.androidMedia.R;
@@ -21,25 +24,29 @@ import static com.harris.androidMedia.MainActivity.TRANSIT_FAB;
 
 public class ExoPlayerMainActivity extends AppCompatActivity implements View.OnClickListener {
 
-    ActivityExoplayerBinding binding;
-
+    Toolbar toolbar;
+    CardView card1;
+    CardView card2;
+    CardView card3;
+    CardView card4;
+    FloatingActionButton fab;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_exoplayer);
-        setSupportActionBar(binding.toolbar);
+        setContentView(R.layout.activity_exoplayer);
+        toolbar = findViewById(R.id.toolbar);
+        toolbar = findViewById(R.id.toolbar);
+        toolbar = findViewById(R.id.toolbar);
+        toolbar = findViewById(R.id.toolbar);
+        fab = findViewById(R.id.fab);
+        setSupportActionBar(toolbar);
 
-        binding.card1.setOnClickListener(this);
-        binding.card2.setOnClickListener(this);
-        binding.card3.setOnClickListener(this);
-        binding.card4.setOnClickListener(this);
-        ViewCompat.setTransitionName(binding.fab, TRANSIT_FAB);
-
-
-
-
-
+        card1.setOnClickListener(this);
+        card2.setOnClickListener(this);
+        card3.setOnClickListener(this);
+        card4.setOnClickListener(this);
+        ViewCompat.setTransitionName(fab, TRANSIT_FAB);
     }
 
     @Override
